refactor: moves all board util related files to a board subfolder
