+++ src/components/Board/Rank/Rank.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
 import Square from '../Square/Square';
-import { files, matchingSquares } from '../../../utils/board';
+import { files, matchingSquares } from '../../../utils/board/board';
 import { RankProps } from './Rank.types';
-import { BoardSquare } from '../../../utils/board.types';
+import { BoardSquare } from '../../../utils/board/board.types';
 
 const Rank: React.FC<RankProps> = (props) => {
   const { number, fullRank, checkedSquare, handlers, data } = props;
