+++ src/components/Board/Rank/Rank.types.ts
@@ -2,7 +2,7 @@ import {
   BoardRank,
   BoardFullRank,
   BoardSquare,
-} from '../../../utils/board.types';
+} from '../../../utils/board/board.types';
 import { BoardData, BoardHandlers } from '../Board/Board.types';
 
 export interface RankProps {
