+++ src/utils/board/board.constants.ts
