+++ src/utils/moves/moves.types.ts
@@ -1,4 +1,4 @@
-import { Board, BoardSquare } from '../board.types';
+import { Board, BoardSquare } from '../board/board.types';
 import { PieceColor } from '../pieces.types';
 
 export type PieceMoveHandler = (
